[core] Don't require peer_ids to match expectations
If a connection has already been made to a particular ip/port and
the peer_id is known, or if a non-compact response is returned by
the tracker so the peer_id for a given remote ip/port is 'known',
do *not* close the connection if the actual client at that ip/port
returns a different peer_id.

While the torrent spec requires that these connections be closed,
there are a few reasons which make this fairly pointless.

1) libtorrent randomises these by default. Other clients probably
randomise these too.

2) It's normal for these to change when a client restarts anyway.
Clients may be restarted for any number of reasons.

3) They're not transmitted by DHT, and are not transmitted by
the tracker by default.

4) The onus is on the *connecting* peer to close the connection,
which means the check is only 50% effective anyway. A client
could still join the swarm by making outgoing connections to
other peers.

If someone really wants this enabled, it can be enabled on a
per-torrent basis using the new setting in TorrentSettings.

Probably address https://github.com/alanmcgovern/monotorrent/issues/627