[PieceWriter] Fix a threading issue with the padding support
Whoops - there was a 'ConfigureAwait(false)' here which caused
improper threaded access to the underlying PieceWriter. The
failure mode was that if there was a torrent with hundreds of
tiny files, multiple threads could be opening/closing streams
at the same time and cause issues due to modifying the underlying
'Streams' dictionary while enumerating it.

i.e.
This exception was originally thrown at this call stack:
    System.ThrowHelper.ThrowArgumentOutOfRange_IndexException()
    System.Collections.Generic.List<T>.RemoveAt(int)
    System.Collections.Generic.List<T>.Remove(T)
    MonoTorrent.PieceWriter.DiskWriter.MaybeRemoveOldestStreams() in DiskWriter.cs
    MonoTorrent.PieceWriter.DiskWriter.GetOrCreateStreamAsync(MonoTorrent.ITorrentManagerFile, System.IO.FileAccess) in DiskWriter.cs
    System.Runtime.CompilerServices.ReusableTaskAwaiter<T>.GetResult()
    MonoTorrent.PieceWriter.DiskWriter.ReadAsync(MonoTorrent.ITorrentManagerFile, long, System.Memory<byte>) in DiskWriter.cs