Use up to 8 concurrent disk writing threads
We can keep up to 196 files open, or whatever the settings allow,
but we should only write to 8 files concurrently.

This is an arbitrary number, but it resolves an issue where occasional
bursts of PieceMessages can result in the same file being opened a few
dozen times as each of the Write requests is launched before any of
them finish.

If the library dropped support for older runtimes, and only supported
,NET 6 or higher, then each FileHandle could be opened exactly once and
many concurrent readers/writers could access the file using the
RandomAcces class. Alas, I'm not fashed about supporting this *yet* in
a backwards compatible way. I just don't want to maintain both
codepaths.