Don't open or truncate files while holding the diskwriter semaphore
These operations can take a (comparatively) substantial amount of time,
so rather than truncating and opening the file immediately, add the
StreamData to the dictionary, and enter it's lock so it's marked as in
use, before swapping to the threadpool to both create and truncate the
file.

This cuts a bunch of time from the IO main loop, making other operations
faster.