Make 'TorrentCreator.StoreSHA1' an optional thing
BEP 47 implies these are required if you want to fully implement that
BEP, but they're not required for standard bittorrent v1 or bittorrent
v2 torrents.

Hashing a file is embarrassingly parallel *except* if BEP47 SHA1 hashes
are computed, in which case it becomes embarassingly linear.

Let's keep things simple and make it optional for whoever is creating
the torrent. Let's not hardcode this to always be on for the small
fraction of people who actually care about it and will use it.