Allow peer messages which are around the size of a bitfield
Torrents with extremely large bitfields can have extremely large
BitfieldMessages. Ensure that we don't discard all incoming connections
if they're simply sending a bitfield for a very large torrent.

An example of this is a 100GB torrent using 16kB pieces.

(Note: This kind of torrent is probably bad for many other reasons, but
why not support it :) )