Fix integration test failures under .net 4.7.2
If a socket is disposed after 'ConnectAsync' is invoked but before the
native 'connect' begins, you can get an 'unexpected'
objectdisposedexceptionw which causes the SocketAsyncEventArgs object to
be left in the 'operation is in progress' state forever.

This doesn't happen under newer frameworks, so use a fallback
implementation for .NET 4.7.2 and netstandard 2.0/2.1.