Fix incorrect assumptions about how the piecelayer hashes are handled
A long time ago I made an assumption that, like elsewhere in the spec,
the final request for piecehashes would be truncated if it extended
beyond the end of the hashes. How wrong I was.

After validating against libtorrent 2.x (using qbittorrent) I confirmed
this doens't happen, and fixed up the impl to no longer truncate the
last message, but to full it with padding hashes instead.

The internal datastructure is still sparse though - it doesn't create
a buffer to hold all the padding hashes.