Fix double-starting a HTTP listener
Now if you call 'Start' twice in a row, the second invocation will be a
no-op as long as the previous call was successful and the listener is
listening.

If the listener failed to bind during the first call, a second attempt
will be made to bind to the specified port.