Tweak prior commit to raise an error instead
Use the original logic to detect when non-sequential requests are made,
and raise an error instead of handling it gracefully.

By default blocks are effectively random, and this is a bad access
pattern for many web servers. The piece picker should be requesting a
bunch of sequential blocks, which should be converted into a single
large http request, which (in turn) should be better for the http server
and also the library.