hash check is now a purely readonly operation
Prior to this hash checking would create empty files, and truncate files
which were too long. However, sometimes people may want to hash check a
file without ever attempting to download it, and in those scenarios
force-truncating it is probably unwise.

Move both the 'Create empty files' logic and 'force-truncate' logic to
StartingMode, a far more apt place. Any time a torrent is started
there's a once-off check to ensure empty files exist (and are marked as
available), and also any files which are too long are truncated.

This is more reasonable, as if someone actually tries to download/seed
a torrent, they should be aware that the files on disk will match what
the torrent requires.