Disambiguate between v1 and v2 connections more clearly
As per BEP52 - if a connection is initiated with the V1 infohash and the relevant reserved bit is set, respond with the V2 infohash to upgrade the connection to 'V2' mode. This makes it unambiguous as to whether or not a given peer supports BEP52.

More importantly, the 'ExpectedInfoHash' property is now the clear indicator as to whether a given peer supports BEP52 or not, as it will always be set to the V2 infohash if the peer supports V2 torrents.

This means peers supporting v1 only can be filtered out when requesting piecelayers from other V2 peers.