Allow resetting ExpectedHash for upgraded v1->v2 connections
When an outgoing connection is being established, only create the
PeerId object after the remote peer has sent it's own handshake.

If the peer responds with a v2 infohash after we send it a v1 infohash,
we do need to create the PeerID object with the new infohash so we treat
it as a v2 connection throughout the library.