Add tests which validate BEPv2 connection upgrading
This test validates that for both incoming and outgoing connections,
both sides of the handshake correctly treat the connection as upgraded
to v2 after a v1 handshake is sent.

This adds tests covering commit
89874d046e345046ed8922be702512aa120b6f55. I over-eagerly merged that
prior to adding the test, but i also reverted the change to ensure this
test failed without those changes.