Fix an inconsistency in handling the partial file suffix
With the new methods in PieceWriter, i discovered a slight inconsistency
in how the suffix was added/removed from partially downloaded files.
While things worked fine for the normal DiskWriter as it was directly
checking on-disk, alternative implementations which relied on
MoveAsync/ExistsAsync/CreateAsync could break.

The issue was that files could have their paths updated prior to the
actual move happening on disk, so 'ITorrentManagerFile.FullPath' would
return the path were we hoped it would be, not where it was currently.

Fixed now!