Fix some racey tests
The leecher is sometimes in 'hashing' mode when the seeder tries to
connect, and so the connections are discarded. Add an explicit wait
to ensure the manager is in an 'active' mode before manually adding any
peers to connect to.

This is only relevant for the tests which strictly control whether the
connection made between the two peers is from seeder -> leecher, or from
leecher -> seeder.

In all other tests there's no race condition as both engines announce to
the tracker, and so whichever one finishes last will be the one which
connects to the first.