Ensure BitTorrent v1 files are in the same order as v2 files
They were out of sync when empty files were in the torrent, however due
to interesting quirks, this didn't cause the infohash or any piece
hashes to be incorrect as only empty files were out-of-order.

Fix this by resorting the files one more time before putting them in the
torrent metadata.

Also - this is only detectable by directly validating the content of the
BEncodedDictionary for some reason. A validation was missing while
parsing the underlying BENcodedDictionary, it was written to operate on
the parsed ITorrentFile objects instead, and at that point it was all
correct.