Split encryption methods into 'tiers'
When establishing an outgoing connection, we may want to try to connect
in multiple ways, for example starting with plain text, and then falling
back to RC4 encryption if that fails. Some clients may want to do
RC4Header first, and then fall back to PlainText and then try RC4Full.

By splitting into tiers, each of these methods can be tried in the
correct order.