Restructure how peer connection retries are executed
The library will now try every supported encryption method each time it
attempts to establish a connection, and it'll try to establish a
connection a fixed number of times before bailing out. Each (overall)
retry will be spaced out in time in case the receiving client is
temporarily busy or inaccessible.