Send the handshake response when connecting to self
If a client connects to itself, send the response handshake before
killing the connection so both sides of the connection know, for sure,
that it is a connection-to-self.

This is a slightly stronger guarantee that the engine won't connect to
itself multiple times in a row before giving up and completely dropping
the connection.