Handle PeerExchangeManager on the right thread
Rather than listening to the events, which are asynchronous, just call
the PeerExchangeManager methods directly when a peer connects or
disconnects. This ensures the methods are called on the main thread.

Probably fixes the issue where null reference exceptions are thrown when
encoding peers in a peer exchange message.

One codepath/thread probably added/removed peers from the list and
another thread did the same. This left things out of sync.