Handle failures of individual dht routers
If a single dht router is offline/unavailable, we should still be able
to successfully resolve the remainder and bootstrap.

Additionally, tweak how initialisation is handled. The list of initial
nodes and the DHT bootstrap nodes must be provided at the start. If both
are provided then the initial nodes will be used for bootstrapping prior
to attempting to resolve/use the DHT bootstrap nodes.

If an empty list of bootstrap nodes is provided, then no lookup will be
performed and only the 'initial nodes' will be used for bootstrapping.

If an empty list of initial nodes and an empty list of bootstrap nodes
are provided, then the engine will fail to initialise.